Mudado para Long o tipo do parâmetro de deletaFoto()

+++ app/src/main/java/com/infact/nightour/controller/FotosController.java
@@ -83,7 +83,7 @@ public class FotosController {
         db.close();
     }
 
-    public void deletaFoto(int id) {
+    public void deletaFoto(Long id) {
         String where = Foto.BD_ID + " = " + id;
 
         db = banco.getWritableDatabase();
